backend/pyproject.toml
[project]
name = "fitness-tracker-backend"
version = "0.2.0"
description = "FastAPI backend for fitness tracker app"
authors = [{ name = "Nicolas Leyton", email = "ndleyton@uc.cl" }]
requires-python = ">=3.10,<3.13"
readme = "README.md"
dependencies = [
    "fastapi>=0.110.0,<0.111",
    "uvicorn[standard]>=0.29.0,<0.31",
    "sqlalchemy>=2.0.29,<3",
    "alembic>=1.13.1,<2",
    "fastapi-users[sqlalchemy, oauth]>=12.1.3,<13",
    "python-dotenv>=1.0.1,<2",
    "asyncpg>=0.29.0,<0.30",
    "passlib[bcrypt]>=1.7.4,<2",
    "pydantic>=2.7.0,<3",
    "pydantic-settings~=2.0",
    "psycopg2-binary>=2.9.10,<3",
    "setuptools>=69.5.1,<70",
    "httpx-oauth>=0.16.1,<0.17",
    "openai>=1.0.0,<2",
    "bcrypt>=4.0.0,<4.1.0",
    "langfuse>=2.0.0,<3",
    "thefuzz>=0.22.1,<0.23",
    "python-levenshtein>=0.25.1,<0.26",
    "google-genai>=1.2.0",
    "pillow>=11.1.0,<12",
]

[dependency-groups]
dev = [
    "pytest>=8.1.1,<9",
    "pytest-cov>=5.0.0,<6",
    "httpx>=0.27.0,<0.28",
    "pytest-asyncio>=1.0.0,<2",
    "pytest-env>=1.0.0,<2",
    "ruff>=0.12.3,<0.13",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
